10370: update the script file for fixing stale automatic blocks
